% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cdx.R
\name{cdx}
\alias{cdx}
\title{Changes directory based on an external reference document}
\usage{
cdx(targ = "list", newPath = FALSE)
}
\arguments{
\item{targ}{A projectory reference id.}

\item{newPath}{If TRUE will prompt to select a folder to assign to targ.}
}
\description{
Changes directory based on an external reference document
}
\examples{
cdx("test",newPath=TRUE)
cdx("test")
}

