% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_table.R
\name{class_table}
\alias{class_table}
\title{Provids a table of variable classes}
\usage{
class_table(dataset)
}
\arguments{
\item{dataset}{a dataset to evaluate}
}
\description{
Provids a table of variable classes
}
\examples{
data(iris)
class_table(iris)
}

