% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/continuous_uniform_time.R
\name{continuous_uniform_time}
\alias{continuous_uniform_time}
\title{Adds non-present time stamps in median increments.}
\usage{
continuous_uniform_time(tmpfile, DateTime)
}
\arguments{
\item{DateTime}{The name of the time variable.}

\item{xtmpfile}{A dataframe to modify.}
}
\description{
Adds non-present time stamps in median increments.
}
\examples{
No examples yet
}

