% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/name_fix.R
\name{name_fix}
\alias{name_fix}
\title{Changes variable names.}
\usage{
name_fix(tmpfile, variable_alias_ref)
}
\arguments{
\item{tmpfile}{A dataframe to modify.}

\item{variable_alias_ref}{A special dataframe which has names to change to.}
}
\description{
Changes variable names.
}
\examples{
No examples yet
}

