% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/omit_NA_cols.R
\name{omit_NA_cols}
\alias{omit_NA_cols}
\title{Removes columns with only missing values.}
\usage{
omit_NA_cols(tmpfile)
}
\arguments{
\item{tmpfile}{A dataframe.}
}
\description{
Removes columns with only missing values.
}
\examples{
a=data.frame(rnorm(10),rep(NA,10),rpois(10,2),c(rep(NA,4),0,0,rep(NA,4)),c(rep(NA,4),1,2,rep(NA,4)),rep(0,10))
print(a)
print(omit_NA_cols(a))
}

