% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/omit_zero_cols.R
\name{omit_zero_cols}
\alias{omit_zero_cols}
\title{Removes columns with only 0 values.}
\usage{
omit_zero_cols(tmpfile)
}
\arguments{
\item{tmpfile}{A dataframe.}
}
\description{
Removes columns with only 0 values.
}
\examples{
a=data.frame(rnorm(10),rep(NA,10),rpois(10,2),c(rep(NA,4),0,0,rep(NA,4)),c(rep(NA,4),1,2,rep(NA,4)),rep(0,10))
print(a)
print(omit_zero_cols(a))
}

