% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbind.all.columns.R
\name{rbind.all.columns}
\alias{rbind.all.columns}
\title{A more robust rbind, allows for different columns.}
\usage{
rbind.all.columns(x, y)
}
\arguments{
\item{x}{A dataframe being appended to.}

\item{y}{A dataframe to append.}
}
\description{
A more robust rbind, allows for different columns.
}

