% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/single_pt_avg_impute.R
\name{single_pt_avg_impute}
\alias{single_pt_avg_impute}
\title{Imputes single values with the mean of pior and post values.
Must be sorted for by time
Does not account for jumps in time, suggest first applying continuous_uniform_time}
\usage{
single_pt_avg_impute(tmpfile)
}
\arguments{
\item{tmpfile}{A data frame to impue}
}
\description{
Imputes single values with the mean of pior and post values.
Must be sorted for by time
Does not account for jumps in time, suggest first applying continuous_uniform_time
}
\examples{
a=data.frame(c(1,2,3,NA,5,6,NA,8,9,10),seq(1,10),c(1,2,NA,NA,5,6,NA,50,9,10),rep("hat",10))
 print(a)
print(single_pt_avg_impute(a))
}

