\name{UslanUtility-package}
\alias{UslanUtility-package}
\alias{UslanUtility}
\docType{package}
\title{
\packageTitle{UslanUtility}
}
\description{
\packageDescription{UslanUtility}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{UslanUtility}
\packageIndices{UslanUtility}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{UslanUtility}

Maintainer: \packageMaintainer{UslanUtility}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
