% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var_range_fix.R
\name{var_range_fix}
\alias{var_range_fix}
\title{Omits specific outlier range.}
\usage{
var_range_fix(tmpfile, var_ranges)
}
\arguments{
\item{tmpfile}{A dataframe to modify.}

\item{var_ranges}{A special dataframe which has ranges to omit.}
}
\description{
Omits specific outlier range.
}
\examples{
No examples yet
}

